requirements.txt
Django>=3.0.6
djangorestframework>=3.11.0
djangorestframework-simplejwt>=4.4.0
psycopg2>=2.8.5
PyJWT>=1.7.1
django-tinymce>=3.0.2

flake8>=3.8.2
